Add time of tours to recycler view

+++ app/src/main/java/com/cs407/team15/redstone/ui/tour/AddTourFragment.kt
@@ -242,7 +242,7 @@ class AddTourFragment : Fragment(){
             }
             distance = Math.floor(distance * 100) / 100
         }
-        val time = (distance*24)
+        val time = (distance*24).toInt()
 
         // create tour object
         val newTour = Tour(name, type, user_id, hammer, locationsOTStr, tagsOnTour, initialVotes, distance)
